fix : 계산 오류 수정
- 더블로 하지 않아서 백분율계산이 되지 않았다

+++ backend/src/main/java/com/codesquad/coco/utils/CalcUtil.java
@@ -6,6 +6,7 @@ public class CalcUtil {
     }
 
     public static int percentCalc(int target, int percent) {
-        return target * (percent / 100);
+        return (int) (target * (percent / 100.0));
     }
+
 }
